Moved presentAssistantMessageContents() to an individual file.
