+++ src/core/cline/clineTypes.ts
@@ -1,0 +1,6 @@
+import { Anthropic } from "@anthropic-ai/sdk"
+
+export type ToolResponse = string | Array<Anthropic.TextBlockParam | Anthropic.ImageBlockParam>
+export type UserContent = Array<
+Anthropic.TextBlockParam | Anthropic.ImageBlockParam | Anthropic.ToolUseBlockParam | Anthropic.ToolResultBlockParam
+>
